Upload arquivos e rotas

+++ server/src/routes/memories.ts
@@ -3,8 +3,14 @@ import { prisma } from "../lib/prisma";
 import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -18,7 +24,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
   });
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -30,6 +36,11 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     });
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
     return memory;
   });
 
@@ -47,13 +58,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "7b51a129-61b3-47e2-8f76-bfc0c5b3fad5",
+        userId: request.user.sub,
       },
     });
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -68,7 +79,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-    const memory = prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId != request.user.sub) {
+      return reply.status(401).send()
+    }
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -78,17 +98,26 @@ export async function memoriesRoutes(app: FastifyInstance) {
         isPublic,
       },
     });
-    
+
     return memory;
   });
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(request.params);
 
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if (memory.userId != request.user.sub) {
+      return reply.status(401).send()
+    }
     await prisma.memory.delete({
       where: {
         id,
